Enable alternative themes for Bootstrap (e.g. Boostwatch)
Alchemy is a great theme, but it was not using the full potential of
underlying technology. Bootstrap provides an easy way to customize look
and feel of rendered pages, and a lot of themes for Bootstrap are freely
available, e.g. at <https://boostwatch.com>

This commit introduces a new configuration variable BOOTSTRAP_CSS to make
use of any existing Bootstrap skins. All incompatibilities in `theme.css`
were fixed:

- Instead of using constant color values we use Bootstrap css variables for
  colors:
  <https://getbootstrap.com/docs/4.3/getting-started/theming/#css-variables>
- The only hardcoded color value left is the color for header/footer
  borders. It was moved into a variable `--alchemy-border` for easier
  overriding if needed. I have tested the current value against both dark
  and light Boostwatch themes, it works ok.
- Pagination template was slightly changed to avoid producing invisible
  text with some of Bootstrap skins