Add configuration parameter to override the stylesheet
This commit adds new configuration parameter, THEME_CSS_OVERRIDES, which
can contain a sequence (list, tuple, etc) of URLs to extra stylesheets
that will be used to override default theme appearance. Both absolute and
relative URLs are supported.

It makes the theme more flexible by allowing small stylesheet tweaks
without requiring to maintain a fork of the whole theme.

When this parameter is not set the theme behaves same as before.